+++ src/components/lookidsmain/TimeCheck.tsx
@@ -1,0 +1,27 @@
+import React, { useEffect } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+import { RootState } from '../../redux/config/configStore';
+import { GET_REWARD, TIME_INTERVAL } from '../../redux/modules/toastSlice';
+
+const TimeCheck = () => {
+    const dispatch = useDispatch();
+    const { page, getReward } = useSelector((state: RootState) => {
+        return state.toast;
+    });
+    useEffect(() => {
+        let timer: any;
+        if (page === 'main') {
+            console.log(getReward);
+            timer = setInterval(() => {
+                dispatch(TIME_INTERVAL());
+            }, 1000);
+        }
+        return () => {
+            clearInterval(timer);
+        };
+    }, [page]);
+
+    return null;
+};
+
+export default TimeCheck;
