+++ src/redux/modules/toastSlice.ts
@@ -29,11 +29,14 @@ const toastSlice = createSlice({
                 state.time -= 1;
             } else {
                 state.getReward = true;
+                state.page = null;
                 state.message = '미션이 완료되었습니다!';
             }
         },
+        GET_REWARD: (state) => {},
         CLEAR_TOAST: (state) => {
             state.message = null;
+            state.page = null;
         },
         CLEAR_TIME: (state) => {
             state.getReward = false;
@@ -43,6 +46,6 @@ const toastSlice = createSlice({
     },
 });
 
-export const { ADD_PAGE, CLEAR_TOAST, TIME_INTERVAL, CLEAR_TIME } = toastSlice.actions;
+export const { ADD_PAGE, CLEAR_TOAST, TIME_INTERVAL, CLEAR_TIME, GET_REWARD } = toastSlice.actions;
 
 export default toastSlice.reducer;
